Decompiler: Escape characters whose Unicode representation is ambiguous
Since the mapping of s2utbl.h is not 1-to-1, utf-8 conversion of the
decompiler output may lose information. For example, both [81 df] and
[87 91] are mapped to U+2261. Pastel Chime uses both of these (!), so
any UTF8 -> SJIS conversion mapping cannot reproduce the original data.

To resolve this issue, this makes the decompiler output these characters
in escaped form like '<0x81df>', when the output encoding is UTF8.