Fix decompilation error in Ouji-sama Lv1
It hit the assertion in analyze_args() because a DATA mark appeared in
the middle of function call sequence. It was added by the
scan_for_data_tables heuristic.

This makes analyze_args() clear such DATA marks, because we are sure
that it is part of an (assignments + funcall) sequence.