ald: Support NL5 manifest format
This adds support for the manifest format used in the System4 SDK's NL5
tool. The format of each line is:

  volume_no, link_no, filename

`ald extract -m` now generates a manifest in this format.

`ald create` still accepts the NL4 manifest format, but the special
notation for non-existent entries (lines with only "0") is no longer
supported.